converted for precision

+++ TriangulatedLocation/src/com/equinix/augumentedReality/Location.java
@@ -9,30 +9,30 @@ package com.equinix.augumentedReality;
  */
 public class Location {
 
-	private int x_cord;
-	private int y_cord;
+	private double x_cord;
+	private double y_cord;
 
 	public Location() {
 
 	}
-	public Location(int x_cord , int y_cord ){
+	public Location(double x_cord , double y_cord ){
 
 		this.x_cord = x_cord;
 		this.y_cord = y_cord;
 	}
-	public int getX_cord() {
+	public double getX_cord() {
 		return x_cord;
 	}
 
-	public void setX_cord(int x_cord) {
+	public void setX_cord(double x_cord) {
 		this.x_cord = x_cord;
 	}
 
-	public int getY_cord() {
+	public double getY_cord() {
 		return y_cord;
 	}
 
-	public void setY_cord(int y_cord) {
+	public void setY_cord(double y_cord) {
 		this.y_cord = y_cord;
 	}
 }
